+++ R/nice_agreement_analysis.R
@@ -1,9 +1,9 @@
 #' Perform within agreement analysis
 #'
-#' This function performs an agreement analysis including intraclass correlation
-#' coefficients, group-related variance, and related diagnostics for items with
-#' a nominal or ordinal scale. The function outputs a data.frame or a HTML table
-#' giving the results.
+#' This function performs an agreement analysis including intra-class
+#' correlation coefficients, group-related variance, and related diagnostics for
+#' items with a nominal or ordinal scale. The function outputs a data.frame or a
+#' HTML table giving the results.
 #'
 #' @param data A data-frame
 #' @param vars Vector of character strings with names of variables to be
@@ -17,23 +17,23 @@
 #' @param n_sim Number of simulations for the agreement analyses.
 #'
 #' @return A Data-frame
-#' 
+#'
 #' @examples
-#' 
-#' vars <- c("Norm_Pop_academic", "Norm_Pop_conform", "Norm_Pop_sporty", 
-#'   "Norm_Pop_cool_clothes", "Norm_Pop_helpful", "Norm_Unpop_academic", 
-#'   "Norm_Unpop_nonconform", "Norm_Unpop_sporty", "Norm_Unpop_cool_clothes", 
-#'   "Norm_Unpop_language", "Norm_LM_academic", "Norm_LM_conform", 
-#'   "Norm_LM_sporty", "Norm_LM_cool_clothes", "Norm_LM_helpful", 
-#'   "Norm_LL_academic", "Norm_LL_nonconform", "Norm_LL_sporty", "Norm_LL_cool_clothes", 
+#'
+#' vars <- c("Norm_Pop_academic", "Norm_Pop_conform", "Norm_Pop_sporty",
+#'   "Norm_Pop_cool_clothes", "Norm_Pop_helpful", "Norm_Unpop_academic",
+#'   "Norm_Unpop_nonconform", "Norm_Unpop_sporty", "Norm_Unpop_cool_clothes",
+#'   "Norm_Unpop_language", "Norm_LM_academic", "Norm_LM_conform",
+#'   "Norm_LM_sporty", "Norm_LM_cool_clothes", "Norm_LM_helpful",
+#'   "Norm_LL_academic", "Norm_LL_nonconform", "Norm_LL_sporty", "Norm_LL_cool_clothes",
 #'   "Norm_LL_language")
-#'   
+#'
 #' nice_agreement_table(
-#'     data = wmisc:::ex_agreement, 
-#'     vars = vars, 
-#'     grouping = "id_class_teacher", 
-#'     rv = 4, 
-#'     crit = 0.6, 
+#'     data = wmisc:::ex_agreement,
+#'     vars = vars,
+#'     grouping = "id_class_teacher",
+#'     rv = 4,
+#'     crit = 0.6,
 #'     n_sim = 100)
 #' @export
 nice_agreement_table <- function(data, 
@@ -51,7 +51,6 @@ nice_agreement_table <- function(data,
 }
 
 #' @export
-#' @rdname nice_agreement_analysis
 agreement_analysis <- function(data, 
                                vars = names(data), 
                                grouping,