+++ NEWS.md
@@ -4,6 +4,7 @@
 ## Changes
 
 - `nice_frequencies()`: Now has a grouping argument. This allows to create contingency tables (e.g. `nice_frequencies(mtcars$cyl, mtcars$am)`).
+- `nice_table()`: New argument `markdown`. If TRUE, interprets cell content as markdown.
 
 # wmisc 0.2.24