+++ NEWS.md
@@ -6,6 +6,7 @@
 - `nice_frequencies()`: Now has a grouping argument. This allows to create contingency tables (e.g. `nice_frequencies(mtcars$cyl, mtcars$am)`).
 - `nice_table()`: New argument `markdown`. If TRUE, interprets cell content as markdown.
 - `ci_score()`: Returns confidence intervalls for testscores.
+- `critical_difference()`
 
 # wmisc 0.2.24